Chat-UI: extract shared attachment helper for chat-input + user-chat
Both components were running line-for-line identical classification +
upload pipelines: classifyFile/classifyDroppedFile, the
ArtifactAttachment shape, and the upload-progress closure that drives
status transitions. Pull them into
`tools/agent-playground/src/lib/components/chat/chat-attachment.ts`:

- `classifyAttachment(file)` — image vs artifact vs unaccepted
- `buildArtifactAttachment(file)` — initial state shape for a fresh drop
- `runArtifactUpload({ att, workspaceId, onUpdate })` — pure helper that
  calls uploadFile and dispatches per-progress / terminal patches via
  an `onUpdate(id, patch)` callback so each consumer owns its $state

`ChatAttachment`, `ImageAttachment`, and `ArtifactAttachment` move to the
helper file; chat-input.svelte re-exports them so existing imports stay
green. The helper is pure TS (no Svelte runes) — directly testable, and
keeps the two consumers from drifting on the next change.
